+++ backend/prisma/sql/fetchHomeTimeline.sql
@@ -5,9 +5,10 @@ SELECT
   p.post_id,
   p.chat_room_id,
   p.user_id,
-  p.content
+  p.content,
+  p.created_at
 FROM post AS p
 WHERE p.user_id = CAST($1 AS UUID)
   AND p.chat_room_id IS NULL
---ORDER BY p.created_at DESC
+ORDER BY p.created_at ASC
 LIMIT $2
