+++ backend/prisma/sql/fetchHomeTimelineCursor.sql
@@ -6,10 +6,11 @@ SELECT
   p.post_id,
   p.chat_room_id,
   p.user_id,
-  p.content
+  p.content,
+  p.created_at
 FROM post AS p
 WHERE p.user_id = CAST($1 AS UUID)
   AND p.chat_room_id IS NULL
   AND CAST($2 AS UUID) = CAST($2 AS UUID)
---ORDER BY p.created_at DESC
+ORDER BY p.created_at ASC
 LIMIT $3
